+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/performance/SyncCommitteePerformance.java
@@ -16,16 +16,17 @@ package tech.pegasys.teku.validator.coordinator.performance;
 import static tech.pegasys.teku.validator.coordinator.performance.DefaultPerformanceTracker.getPercentage;
 
 import java.util.Objects;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
 public class SyncCommitteePerformance {
-  private final int epoch;
+  private final UInt64 epoch;
   private final int numberOfExpectedMessages;
   private final int numberOfProducedMessages;
   private final int numberOfCorrectMessages;
   private final int numberOfIncludedMessages;
 
   public SyncCommitteePerformance(
-      final int epoch,
+      final UInt64 epoch,
       final int numberOfExpectedMessages,
       final int numberOfProducedMessages,
       final int numberOfCorrectMessages,
@@ -62,11 +63,11 @@ public class SyncCommitteePerformance {
       return false;
     }
     final SyncCommitteePerformance that = (SyncCommitteePerformance) o;
-    return epoch == that.epoch
-        && numberOfExpectedMessages == that.numberOfExpectedMessages
+    return numberOfExpectedMessages == that.numberOfExpectedMessages
         && numberOfProducedMessages == that.numberOfProducedMessages
         && numberOfCorrectMessages == that.numberOfCorrectMessages
-        && numberOfIncludedMessages == that.numberOfIncludedMessages;
+        && numberOfIncludedMessages == that.numberOfIncludedMessages
+        && epoch.equals(that.epoch);
   }
 
   @Override
@@ -83,7 +84,7 @@ public class SyncCommitteePerformance {
   public String toString() {
     return String.format(
         "Sync committee performance: "
-            + "epoch %d, expected %d, produced %d, correct %d, included %d (%d%%)",
+            + "epoch %s, expected %s, produced %s, correct %s, included %s (%s%%)",
         epoch,
         numberOfExpectedMessages,
         numberOfProducedMessages,