+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/performance/SyncCommitteePerformanceTracker.java
@@ -176,7 +176,7 @@ public class SyncCommitteePerformanceTracker {
         .thenApply(
             numberOfIncludedMessages ->
                 new SyncCommitteePerformance(
-                    epoch.intValue(),
+                    epoch,
                     numberOfExpectedMessages,
                     numberOfProducedMessages,
                     numberOfCorrectMessages,